bot/requirements.txt
python-telegram-bot>=20,<22
python-dotenv>=1,<3
pyodbc>=4,<5

# ML
tensorflow-cpu==2.19.0   # o tensorflow==2.19.0 si necesitas GPU; no fijes 'keras'

# NumPy/Pandas/Sklearn compatibles con TF 2.19
numpy>=1.26,<2.0
pandas>=1.5,<2.0
scikit-learn>=1.3,<1.5

# Utilidades
requests>=2.25,<3.0
reportlab==4.4.1
Pillow>=10,<11
openpyxl>=3.1,<4.0